Split storage into multiple files.
